refactor: move get_shard_iterator to helper_functions
