+++ tests/aws/services/kinesis/helper_functions.py
@@ -1,0 +1,16 @@
+def get_shard_iterator(stream_name, kinesis_client):
+    response = kinesis_client.describe_stream(StreamName=stream_name)
+    sequence_number = (
+        response.get("StreamDescription")
+        .get("Shards")[0]
+        .get("SequenceNumberRange")
+        .get("StartingSequenceNumber")
+    )
+    shard_id = response.get("StreamDescription").get("Shards")[0].get("ShardId")
+    response = kinesis_client.get_shard_iterator(
+        StreamName=stream_name,
+        ShardId=shard_id,
+        ShardIteratorType="AT_SEQUENCE_NUMBER",
+        StartingSequenceNumber=sequence_number,
+    )
+    return response.get("ShardIterator")
