+++ tests/aws/services/kinesis/test_kinesis.py
@@ -20,30 +20,13 @@ from localstack.testing.pytest import markers
 from localstack.utils.aws import resources
 from localstack.utils.common import retry, select_attributes, short_uid
 from localstack.utils.kinesis import kinesis_connector
+from tests.aws.services.kinesis.helper_functions import get_shard_iterator
 
 LOGGER = logging.getLogger(__name__)
 
 # make sure cbor2 patches are applied
 # (for the test-data decoding, usually done as init hook in LocalStack)
 _patch_cbor2()
-
-
-def get_shard_iterator(stream_name, kinesis_client):
-    response = kinesis_client.describe_stream(StreamName=stream_name)
-    sequence_number = (
-        response.get("StreamDescription")
-        .get("Shards")[0]
-        .get("SequenceNumberRange")
-        .get("StartingSequenceNumber")
-    )
-    shard_id = response.get("StreamDescription").get("Shards")[0].get("ShardId")
-    response = kinesis_client.get_shard_iterator(
-        StreamName=stream_name,
-        ShardId=shard_id,
-        ShardIteratorType="AT_SEQUENCE_NUMBER",
-        StartingSequenceNumber=sequence_number,
-    )
-    return response.get("ShardIterator")
 
 
 class TestKinesis:
